small fix for cron
git-svn-id: http://svn.omploader.org/trunk/omploader@256 f3e9fb10-4bd1-4be2-8a77-beb9a613051d

+++ scripts/omploader.rb
@@ -114,7 +114,7 @@ def run_cron(db)
 	num_rows = stmt.num_rows
 	num_rows.times do
 		id = stmt.fetch.to_s
-		File.unlink(Paths['thumbnails'] + '/' + id.to_b64)
+		File.unlink(Paths['thumbnails'] + '/' + id.to_b64) if File.exist?(Paths['thumbnails'] + '/' + id.to_b64)
 	end
 	stmt.close
 	q = db.prepare('update metadata inner join thumbnails on thumbnails.id = metadata.thumbnail_id set metadata.thumbnail_id = null where unix_timestamp(thumbnails.last_accessed) < unix_timestamp(current_timestamp) - ?')
